refactor: react-query 사용 및 폴더 구조 변경 (#44)
