+++ src/hooks/queries/useFetchStrategyDetail.ts
